Restructed code and updated zip code logic
