+++ lib/pages/login/phone_sign_in_screen.dart
@@ -4,7 +4,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 
-import 'navbar_setup.dart';
+import '../navbar_setup.dart';
 
 class PhoneSignInScreen extends StatefulWidget {
   @override
