+++ pubspec.yaml
@@ -41,6 +41,8 @@ dependencies:
   cloud_firestore: ^5.4.5
   google_maps_flutter: ^2.1.0
   geolocator: ^9.0.2
+  flutter_google_places: ^0.3.0
+  google_maps_webservice: ^0.0.20-nullsafety.5
 
 dev_dependencies:
   flutter_test:
